requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.2.1
google-genai==1.65.0
requests==2.32.5
beautifulsoup4==4.14.3
PyMuPDF==1.27.1
Pillow==12.1.1
waitress==3.0.2
